requirements.txt
google-generativeai==0.7.2
vaderSentiment==3.3.2
python-dotenv==1.0.1
cryptography==43.0.0
gtts==2.5.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
